+++ internal/ceres/program.cc
@@ -439,7 +439,8 @@ Program::CreateJacobianBlockSparsityTranspose(int start_residual_block) const {
   const int num_cols = NumResidualBlocks() - start_residual_block;
 
   std::unique_ptr<TripletSparseMatrix> tsm(
-      new TripletSparseMatrix(num_rows, num_cols, 10 * num_cols));
+      new TripletSparseMatrix(num_rows, num_cols,
+                              10 * static_cast<int64_t>(num_cols)));
   int64_t num_nonzeros = 0;
   int* rows = tsm->mutable_rows();
   int* cols = tsm->mutable_cols();
